checkpoint: renames "Server/ServerConnectionError.ts" to "Server/Error/Connection.ts"
