+++ src/features/TextToImage/Server/Error/Connection.ts
